# 1.11 Fix patching virtualOutbound-blackhole (#36329) (#36366)