--- a/releasenotes/notes/envoyfilter-patch-duplicate.yaml
+++ b/releasenotes/notes/envoyfilter-patch-duplicate.yaml
@@ -1,0 +1,8 @@
+apiVersion: release-notes/v2
+kind: bug-fix
+area: traffic-management
+issues:
+- 35778
+releaseNotes:
+- |
+  **Fixed** an issue where `EnvoyFilter` patches on `virtualOutbound-blackhole` could cause memory leaks.